// src/app/interfaces/IStudentTc.ts
export interface IStudentTc {
  tcId: number;
  stuId: string;
  rollNumber?: string | null;
  classSectionId: number;
  houseId?: number;
  categoryId?: number;
  anyOtherRemark?: string;
  generalContact?: string;
  reasonLevingSchool?: string;
  dateOfApplication?: string;
  dateOfIssue?: string;
  extraActivity?: string;
  nccCadet?: string;
  concession?: string;
  failSameClass?: string;
  boardAnnualExaminationLast?: string;
  bookNo?: string;
  slNo?: string;
  admissionNo?: string;
  ssmid?: string;
  studentName?: string;
  fatherName?: string;
  motherName?: string;
  dobFigure?: string;
  dobWord?: string;
  nationality?: string;
  category?: string;
  dateOfAdmisson?: string;
  classStudies?: string;
  subjectStudies?: string;
  pramotedToFigure?: string;
  pramotedToWord?: string;
  paidSchoolDues?: string;
  boardName?: string;
  session?: number;
  workingDays?: number;
  presentDays?: number;
  communicationAddress?: string;
  penNo?: string;
}